Rename files to match functions
